buildroot-external-lima: create repository
This is a BR2_EXTERNAL tree for Buildroot which contains modifications
to the mesa3d package so that it supports building the mesa port of
lima, to be used for people interested in developing or trying lima.

It contains some hacks since we need to modify the existing mesa3d
Buildroot upstream recipe, and that is not very well supported in
Buildroot. Therefore, this was done in the seemingly least terrible way.
Still, it should be enough for lima development while we can't merge
everything to their respective upstream projects yet, and all of this
repo should become unnecessary when that happens.

Signed-off-by: Erico Nunes <nunes.erico@gmail.com>